backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# passlib 1.7.4's compat probe trips bcrypt 4.1's new 72-byte rejection,
# breaking every hash/verify call. Pin until passlib ships a fix.
bcrypt<4.1
python-multipart>=0.0.9
anthropic>=0.25.0
boto3>=1.34.0
pytesseract>=0.3.10
slowapi>=0.1.9
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
ruff>=0.3.0
mypy>=1.9.0
